+++ src/FunctionApp8.0/Program.cs
@@ -1,3 +1,13 @@
+using System.Security.Claims;
+using CleanArchitecture.Presentation.FunctionApp8;
+using CleanArchitecture8.Application.Common.Interfaces;
+using Microsoft.AspNetCore.Http;
+using Microsoft.Azure.Functions.Worker.Extensions.OpenApi.Extensions;
+using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Abstractions;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+
 var hostBuilder = new HostBuilder()
      .ConfigureFunctionsWorkerDefaults(
     worker =>
@@ -19,12 +29,11 @@ hostBuilder.ConfigureServices((context, services) =>
 {
     var configuration = context.Configuration;
 
-    services.AddSingleton<IHttpRequestProcessor, HttpRequestProcessor>();
     services.AddScoped<IUser, CurrentUser>();
 
     services.AddApplicationServices();
     services.AddInfrastructureServices(configuration);
-    services.AddWebServices();
+    //services.AddWebServices();
 
     //// set OpenAPIOptions for Swagger support
     var version = configuration.GetValue<string>("Version") ?? "1.0";
@@ -40,7 +49,9 @@ var host = hostBuilder.Build();
 
 await host.RunAsync();
 
-//internal class CurrentUserService : ICurrentUserService
+/// <summary>
+/// Used in Bihaviours
+/// </summary>
 public class CurrentUser : IUser
 {
     private readonly IHttpContextAccessor _httpContextAccessor;
@@ -51,29 +62,4 @@ public class CurrentUser : IUser
     }
 
     public string? Id => _httpContextAccessor.HttpContext?.User?.FindFirstValue(ClaimTypes.NameIdentifier);
-}
-public class HttpOpenApiConfigurationOptions : OpenApiConfigurationOptions
-{
-    private readonly string _version;
-
-    public HttpOpenApiConfigurationOptions(string version)
-    {
-        _version = version;
-    }
-        /// <summary>
-    /// Include host name.
-    /// </summary>
-    public override bool IncludeRequestingHostName { get { return false; } }
- 
-    /// <summary>
-    /// OpenAPI version.
-    /// </summary>
-    public override OpenApiVersionType OpenApiVersion { get { return OpenApiVersionType.V3; } }
-
-    public override OpenApiInfo Info { get; set; } = new OpenApiInfo
-    {
-        Title = "My Azure Function API",
-        Version = "1.0.0",  // Default value (overridden later)
-        Description = "API documentation for the Azure Functions project with Clean Architecture."
-    };
 }
